Isolate the DRF integration
This improves import-time performance by a ton, which is especially
useful for the Django development workflow.
